+++ src/components/RightSidebar.vue
@@ -9,6 +9,7 @@
       <OptimizationSection />
       <ExportSection />
       <FieldListSection />
+      <Authenticated />
     </div>
   </aside>
 </template>
@@ -22,6 +23,7 @@ import DipulCheckSection from "./sidebar/DipulCheckSection.vue"
 import OptimizationSection from "./sidebar/OptimizationSection.vue"
 import ExportSection from "./sidebar/ExportSection.vue"
 import FieldListSection from "./sidebar/FieldListSection.vue"
+import Authenticated from "./sidebar/Authenticated.vue";
 
 onMounted(() => {
   console.log("RightSidebar mounted")