+++ vite.config.ts
@@ -6,4 +6,7 @@ import tailwindcss from '@tailwindcss/vite'
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [vue(),  tailwindcss()],
+    server: {
+      allowedHosts: ["ag-fe.bury.local","agmap.bury.local","ag.bury.link", "ag.bury.local"]
+    }
 })